requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==2.3.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==2.3.5
openai==0.27.8
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==25.0
pillow==10.4.0
polars==1.38.1
polars-runtime-32==1.38.1
proglog==0.1.12
propcache==0.4.1
psutil==7.2.2
pyparsing==3.3.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.31.0
safetensors==0.7.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
tokenizers==0.22.1
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==2.3.7
yarl==1.22.0
